+++ Tests/Unit/Controller/UserWithoutAutologinControllerTest.php
@@ -4,7 +4,12 @@ declare(strict_types=1);
 
 namespace OliverKlee\Onetimeaccount\Tests\Unit\Controller;
 
+use OliverKlee\FeUserExtraFields\Domain\Model\FrontendUser;
 use OliverKlee\Onetimeaccount\Controller\UserWithoutAutologinController;
+use Prophecy\Argument;
+use Prophecy\Prophecy\ObjectProphecy;
+use TYPO3\CMS\Frontend\Authentication\FrontendUserAuthentication;
+use TYPO3\CMS\Frontend\Controller\TypoScriptFrontendController;
 
 /**
  * @extends AbstractUserControllerTest<UserWithoutAutologinController>
@@ -14,10 +19,19 @@ use OliverKlee\Onetimeaccount\Controller\UserWithoutAutologinController;
  */
 final class UserWithoutAutologinControllerTest extends AbstractUserControllerTest
 {
+    /**
+     * @var ObjectProphecy<FrontendUserAuthentication>
+     *
+     * We can make this property private once we drop support for TYPO3 V9.
+     */
+    protected $userProphecy;
+
     protected function setUp(): void
     {
         parent::setUp();
         $this->setDummyRequestData();
+
+        $this->setUpFakeFrontEnd();
 
         // We need to create an accessible mock in order to be able to set the protected `view`.
         // We can drop the additional arguments to skip the original constructor once we drop support for TYPO3 V9.
@@ -30,5 +44,38 @@ final class UserWithoutAutologinControllerTest extends AbstractUserControllerTes
         );
 
         $this->setUpAndInjectSharedDependencies();
+    }
+
+    protected function tearDown(): void
+    {
+        unset($GLOBALS['TSFE']);
+        parent::tearDown();
+    }
+
+    private function setUpFakeFrontEnd(): void
+    {
+        $this->userProphecy = $this->prophesize(FrontendUserAuthentication::class);
+        $user = $this->userProphecy->reveal();
+
+        $frontEndController = $this->prophesize(TypoScriptFrontendController::class)->reveal();
+        $frontEndController->fe_user = $user;
+        $GLOBALS['TSFE'] = $frontEndController;
+    }
+
+    /**
+     * @test
+     */
+    public function storesUidOfNewUserInSession(): void
+    {
+        $userUid = 42;
+        $user = new FrontendUser();
+        $user->_setProperty('uid', $userUid);
+
+        $this->credentialsGeneratorProphecy->generateUsernameForUser(Argument::any());
+        $this->credentialsGeneratorProphecy->generatePasswordForUser(Argument::any())->willReturn('hashed-password');
+
+        $this->userProphecy->setAndSaveSessionData('onetimeaccountUserUid', $userUid)->shouldBeCalled();
+
+        $this->subject->createAction($user);
     }
 }